HSD #15016141215: i2c: designware: Fix SOCFPGA-specific recovery issue in MODEL_MASK (#221)
After rebase to kernel 6.6.x, the MODEL_MASK macro did not include
BIT 12, which is now assigned to MODEL_SOCFPGA due to the addition
of a new model (WANGXUN_SP) that uses BIT 11. As a result, the
SOCFPGA-specific I2C recovery function was not invoked, leading to
issues such as the SCL line being stuck in some cases.

Signed-off-by: Tanmay Kathpalia <tanmay.kathpalia@intel.com>