HSD#15017576547: arm64:dts: agilex: Disable sdm interrupt support
fcs in agilex uses svc V1 and sends a poll service status
command to get the response for random number. Before
polling, an interrupt is triggered and the response buffer
is not getting properly populated in ATF when svc
performs service complete smc call.

Disabling an interrupt so the response buffer is properly
populated by the atf when poll service status command is
issued by fcs. Interrupt will be enabled once the
svc client, FCS, moves to svc V3

Signed-off-by: Balsundar Ponnusamy <balsundar.ponnusamy@intel.com>