HSD #16026788956: net: stmmac: fpe: Keep pmac always enabled
The preemptive mac in DWMAC is always enabled, that means
DWMAC is ready to receive preemptive mPackets always. So
keep pmac_enabled true. pmac_enabled = false is not valid
for this IP.

Also, FPE residual queue is the Rx Queue to which residual
preemptive mPackets must be forwarded from the pmac. So
moved the residual queue configuration to the pmac config
instead of tx config.

Signed-off-by: Rohan G Thomas <rohan.g.thomas@altera.com>

State: pending